+++ src/main/java/io/contek/invoker/binancespot/api/common/_BorrowRate.java
@@ -1,7 +1,6 @@
 package io.contek.invoker.binancespot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 import java.util.List;
 
 @NotThreadSafe
@@ -11,8 +10,8 @@ public class _BorrowRate {
   public String coin;
   public boolean transferIn;
   public boolean borrowable;
-  public BigDecimal dailyInterest;
-  public BigDecimal yearlyInterest;
-  public BigDecimal borrowLimit;
+  public Double dailyInterest;
+  public Double yearlyInterest;
+  public Double borrowLimit;
   public List<String> marginablePairs;
 }
