+++ src/main/java/io/contek/invoker/binancespot/api/common/_CrossCollateral.java
@@ -1,17 +1,16 @@
 package io.contek.invoker.binancespot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _CrossCollateral {
 
   public String loanCoin;
   public String collateralCoin;
-  public BigDecimal locked;
-  public BigDecimal loanAmount;
-  public BigDecimal currentCollateralRate;
-  public BigDecimal interestFreeLimitUsed;
-  public BigDecimal principalForInterest;
-  public BigDecimal interest;
+  public Double locked;
+  public Double loanAmount;
+  public Double currentCollateralRate;
+  public Double interestFreeLimitUsed;
+  public Double principalForInterest;
+  public Double interest;
 }
