+++ src/main/java/io/contek/invoker/binancespot/api/common/_MarginAccount.java
@@ -1,17 +1,16 @@
 package io.contek.invoker.binancespot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 import java.util.List;
 
 @NotThreadSafe
 public class _MarginAccount {
 
   public boolean borrowEnabled;
-  public BigDecimal marginLevel;
-  public BigDecimal totalAssetOfBtc;
-  public BigDecimal totalLiabilityOfBtc;
-  public BigDecimal totalNetAssetOfBtc;
+  public Double marginLevel;
+  public Double totalAssetOfBtc;
+  public Double totalLiabilityOfBtc;
+  public Double totalNetAssetOfBtc;
   public boolean tradeEnabled;
   public boolean transferEnabled;
   public List<_MarginAsset> userAssets;
