+++ src/main/java/io/contek/invoker/binancespot/api/common/_MarginAccountSummary.java
@@ -1,13 +1,12 @@
 package io.contek.invoker.binancespot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _MarginAccountSummary {
 
-  public BigDecimal uniMMR; // Portfolio margin account maintenance margin rate
-  public BigDecimal accountEquity; // Account equity, unit：USD
-  public BigDecimal accountMaintMargin; // Portfolio margin account maintenance margin, unit：USD
+  public Double uniMMR; // Portfolio margin account maintenance margin rate
+  public Double accountEquity; // Account equity, unit：USD
+  public Double accountMaintMargin; // Portfolio margin account maintenance margin, unit：USD
   public String accountStatus; // Portfolio margin account status:"NORMAL", "MARGIN_CALL", "SUPPLY_MARGIN", "REDUCE_ONLY", "ACTIVE_LIQUIDATION", "FORCE_LIQUIDATION", "BANKRUPTED"
 }
