+++ src/main/java/io/contek/invoker/binancespot/api/common/_PriceTicker.java
@@ -1,11 +1,10 @@
 package io.contek.invoker.binancespot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _PriceTicker {
 
   public String symbol;
-  public BigDecimal price;
+  public Double price;
 }
